% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_function.R
\name{add}
\alias{add}
\title{Add together two numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{A number that is being added.}

\item{y}{A number that is being added.}
}
\value{
A number obtained in the sum process.
}
\description{
Add together two numbers
}
\examples{
add(10, 2.5)
add(8, 3)
}
