% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{count_records}
\alias{count_records}
\title{Count the Records of a Table in a Remote DB}
\usage{
count_records(table_name, conn)
}
\arguments{
\item{table_name}{The name of the table to count the records of}

\item{conn}{A DBI connection object}
}
\description{
Count the Records of a Table in a Remote DB
}
