% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_function.R
\name{divide_by}
\alias{divide_by}
\title{Division of two numbers}
\usage{
divide_by(dividend, divisor)
}
\arguments{
\item{dividend}{A number that is being divided in the division process.}

\item{divisor}{A number that by which the dividend is being divided by.}
}
\value{
A number obtained in the division process.
}
\description{
Division of two numbers
}
\examples{
divide_by(10, 2.5)
divide_by(8, 3)
}
