% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_function.R
\name{nested_function}
\alias{nested_function}
\title{Nested function}
\usage{
nested_function(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
A number obtained using both add and divide_by functions.
}
\description{
Nested function
}
\examples{
nested_function(10, 2)
}
