% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_dots}
\alias{plot_dots}
\title{Plot Some Dots}
\usage{
plot_dots(n_dots = 100L, ...)
}
\arguments{
\item{n_dots}{Number of dots to plot}

\item{...}{Additional argument passed on to \code{ggplot2::geom_point}}
}
\description{
Plot Some Dots
}
